refactor(sim): decouple sim state and pickup event type roots
